requirements.txt
langchain-google-genai
python-dotenv
Flask==2.2.5
flask-cors==3.0.10
mysql-connector-python==8.0.33
google-generativeai
SpeechRecognition
gTTS
sounddevice
scipy
